--- src/components/searchBox/searchBox.css
@@ -1,17 +1,0 @@
-.MuiInputBase-root {
-  background-color: #ffffff !important;
-}
-
-input:hover, input:active, input:focus {
-  border: none !important;
-  outline: none !important;
-  box-shadow: none !important;
-} */
-
-svg {
-  margin-right: .8em;
-}
-
-/* .main {
-  background-color: #202029;
-} */
